made most things modular
