+++ config/GoogleMapsEmbed.tsx
@@ -1,0 +1,11 @@
+export default function GoogleMapsEmbed() {
+  return (
+    <iframe
+      src="https://storage.googleapis.com/maps-solutions-hkl2kro3e4/locator-plus/gprp/locator-plus.html"
+      width="100%"
+      height="100%"
+      style={{ border: 0 }}
+      loading="lazy"
+    ></iframe>
+  )
+}
