+++ config/carriers.ts
@@ -1,0 +1,74 @@
+import { Carrier } from "@/types"
+import { Metadata } from "next"
+
+export const carriersMetaData: Metadata = {
+  title: "",
+  description: "",
+  keywords: [],
+}
+
+export const carriers: Carrier[] = [
+  {
+    name: "Acme Insurance",
+    description: "",
+    descriptionShort:
+      "Providing comprehensive coverage for individuals and businesses.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+  {
+    name: "Apex Insurance",
+    description: "",
+    descriptionShort:
+      "Specializing in auto, home, and life insurance solutions.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+  {
+    name: "Zenith Insurance",
+    description: "",
+    descriptionShort:
+      "Offering tailored insurance solutions for businesses and individuals.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+  {
+    name: "Emerald Insurance",
+    description: "",
+    descriptionShort:
+      "Providing innovative insurance products for modern lifestyles.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+  {
+    name: "Horizon Insurance",
+    description: "",
+    descriptionShort:
+      "Protecting your future with reliable and affordable coverage.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+  {
+    name: "Pinnacle Insurance",
+    description: "",
+    descriptionShort: "Customizing insurance plans to fit your unique needs.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+  {
+    name: "Evergreen Insurance",
+    description: "",
+    descriptionShort:
+      "Committed to delivering exceptional service and peace of mind.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+  {
+    name: "Sapphire Insurance",
+    description: "",
+    descriptionShort:
+      "Empowering you to make informed decisions about your coverage.",
+    logo: "/placeholder-logo.svg",
+    url: "",
+  },
+]
