+++ config/faq.ts
@@ -52,275 +52,3 @@ export const defaultFAQs: FAQ[] = [
       "It's a good idea to review your insurance coverage annually or whenever you experience significant life changes, such as buying a new home, getting married, or having a child. Regular reviews ensure your coverage remains adequate and up-to-date.",
   },
 ]
-
-export const homeownersFAQ: FAQ[] = [
-  {
-    question: "What is covered under a standard homeowners insurance policy?",
-    answer:
-      "A standard homeowners insurance policy covers your home, personal belongings, and additional living expenses in the event of a covered peril such as fire, theft, or certain natural disasters. It also includes liability protection for injuries to others on your property.",
-  },
-  {
-    question: "How is the cost of homeowners insurance determined?",
-    answer:
-      "The cost of homeowners insurance is determined by factors such as the value of your home, location, coverage limits, deductible, age and condition of your home, claims history, and sometimes your credit score.",
-  },
-  {
-    question: "Is flood damage covered by homeowners insurance?",
-    answer:
-      "No, standard homeowners insurance policies do not cover flood damage. You will need to purchase a separate flood insurance policy if you live in a flood-prone area.",
-  },
-  {
-    question: "What is liability coverage in homeowners insurance?",
-    answer:
-      "Liability coverage in homeowners insurance protects you from financial loss if you are found legally responsible for someone else's injury or property damage that occurs on your property.",
-  },
-  {
-    question: "What should I do if my home is damaged by a covered peril?",
-    answer:
-      "If your home is damaged by a covered peril, contact your insurance provider to file a claim, document the damage with photos, and keep receipts for any expenses related to temporary repairs or additional living costs.",
-  },
-  {
-    question:
-      "Can I bundle homeowners insurance with other types of insurance?",
-    answer:
-      "Yes, many insurance providers offer discounts for bundling homeowners insurance with other policies such as auto or life insurance. Bundling can lead to significant savings on your premiums.",
-  },
-]
-
-export const commercialFAQ: FAQ[] = [
-  {
-    question: "What does commercial insurance cover?",
-    answer:
-      "Commercial insurance covers risks for businesses, including property damage, liability claims, employee injuries, and business interruption. Policies can be customized to meet the specific needs of your business.",
-  },
-  {
-    question: "How much does commercial insurance cost?",
-    answer:
-      "The cost of commercial insurance varies based on factors such as industry, location, coverage limits, and number of employees. It can range from a few hundred to several thousand dollars annually.",
-  },
-  {
-    question: "Why is business interruption insurance important?",
-    answer:
-      "Business interruption insurance provides coverage for lost income and operating expenses if your business operations are disrupted due to a covered event, ensuring financial stability during unexpected closures.",
-  },
-  {
-    question: "What is general liability insurance for businesses?",
-    answer:
-      "General liability insurance protects businesses against claims of bodily injury, property damage, and personal injury arising from business operations. It covers legal costs and damages awarded in lawsuits.",
-  },
-  {
-    question: "Can I customize my commercial insurance policy?",
-    answer:
-      "Yes, commercial insurance policies can be customized to meet the specific needs of your business, including property coverage, liability coverage, workers' compensation, and business interruption insurance.",
-  },
-  {
-    question:
-      "What is the difference between commercial property insurance and general liability insurance?",
-    answer:
-      "Commercial property insurance covers damage to your business property, while general liability insurance protects against claims of bodily injury and property damage to third parties resulting from your business operations.",
-  },
-]
-
-export const autoFAQ: FAQ[] = [
-  {
-    question: "What types of auto insurance coverage are available?",
-    answer:
-      "Auto insurance coverage includes liability coverage, collision coverage, comprehensive coverage, personal injury protection, and uninsured/underinsured motorist coverage.",
-  },
-  {
-    question: "How can I lower my auto insurance premiums?",
-    answer:
-      "You can lower your auto insurance premiums by maintaining a clean driving record, bundling your auto insurance with other policies, increasing your deductible, and taking advantage of available discounts.",
-  },
-  {
-    question:
-      "What is the difference between collision and comprehensive coverage?",
-    answer:
-      "Collision coverage pays for damages to your vehicle resulting from a collision with another car or object, while comprehensive coverage protects your vehicle from non-collision-related incidents such as theft, vandalism, and natural disasters.",
-  },
-  {
-    question: "Do I need uninsured motorist coverage?",
-    answer:
-      "Uninsured motorist coverage is recommended as it protects you in the event of an accident with an uninsured or underinsured driver, covering your medical expenses and vehicle repairs.",
-  },
-  {
-    question: "What factors affect my auto insurance premiums?",
-    answer:
-      "Factors that affect auto insurance premiums include your driving record, age, type of vehicle, coverage limits, deductible, location, and sometimes your credit history.",
-  },
-  {
-    question: "How do I file an auto insurance claim?",
-    answer:
-      "To file an auto insurance claim, contact your insurance provider as soon as possible, document the damage with photos, gather relevant information such as the police report and other driver's details, and follow your insurer's claims process.",
-  },
-]
-
-export const rentersFAQ: FAQ[] = [
-  {
-    question: "What does renters insurance cover?",
-    answer:
-      "Renters insurance covers your personal belongings against risks such as theft, fire, and water damage. It also provides liability protection and additional living expenses if your rental unit becomes uninhabitable due to a covered event.",
-  },
-  {
-    question: "How much does renters insurance cost?",
-    answer:
-      "The cost of renters insurance typically ranges from $10 to $30 per month, depending on the coverage limits, deductible, and location of the rental property.",
-  },
-  {
-    question: "Is renters insurance mandatory?",
-    answer:
-      "While renters insurance is not legally required, many landlords require tenants to have a policy as part of the lease agreement to protect against potential damages and liabilities.",
-  },
-  {
-    question: "Can renters insurance cover roommates?",
-    answer:
-      "Renters insurance policies generally cover only the named insured and their family members. Roommates should each have their own renters insurance policy to ensure adequate coverage for their belongings.",
-  },
-  {
-    question: "What are the benefits of renters insurance?",
-    answer:
-      "Renters insurance provides financial protection for your belongings, liability coverage for accidents in your rental unit, and additional living expenses if you need to temporarily relocate due to a covered event.",
-  },
-  {
-    question: "How do I file a renters insurance claim?",
-    answer:
-      "To file a renters insurance claim, contact your insurance provider as soon as possible, document the damage or loss, provide necessary receipts or evidence, and follow your insurer's claims process.",
-  },
-]
-
-export const landlordFAQ: FAQ[] = [
-  {
-    question: "What does landlord insurance cover?",
-    answer:
-      "Landlord insurance covers property damage, liability claims, and loss of rental income for property owners renting out their properties. It also provides coverage for landlord-owned items within the rental property, such as appliances and furniture.",
-  },
-  {
-    question: "How much does landlord insurance cost?",
-    answer:
-      "The cost of landlord insurance varies based on factors such as the value of the property, location, coverage limits, and deductible. On average, landlords can expect to pay between $800 and $1,500 annually.",
-  },
-  {
-    question: "Do I need landlord insurance if I rent out my property?",
-    answer:
-      "Yes, landlord insurance is recommended for property owners who rent out their properties. It provides essential coverage for property damage, liability claims, and loss of rental income, ensuring your investment is protected.",
-  },
-  {
-    question:
-      "What is the difference between landlord insurance and homeowners insurance?",
-    answer:
-      "Landlord insurance covers rental properties and provides additional protections for rental income and landlord liabilities, while homeowners insurance covers owner-occupied properties. Landlord insurance is necessary if you rent out your property.",
-  },
-  {
-    question: "Can landlord insurance cover tenant damage?",
-    answer:
-      "Yes, landlord insurance can cover damage caused by tenants, depending on the policy. It typically covers accidental damage, but intentional damage by tenants may not be covered.",
-  },
-  {
-    question: "What is loss of rental income coverage?",
-    answer:
-      "Loss of rental income coverage helps compensate for lost rental income if your property becomes uninhabitable due to a covered event. It ensures that you continue to receive rental income while repairs are being made.",
-  },
-]
-
-export const lifeFAQ: FAQ[] = [
-  {
-    question: "What types of life insurance are available?",
-    answer:
-      "Life insurance options include term life insurance, whole life insurance, universal life insurance, and variable life insurance. Each type offers different benefits and features to meet your financial needs.",
-  },
-  {
-    question: "How much does life insurance cost?",
-    answer:
-      "The cost of life insurance depends on factors such as age, health, coverage amount, and policy type. On average, term life insurance can cost between $20 and $50 per month, while whole life insurance can range from $100 to $300 per month.",
-  },
-  {
-    question: "Why is life insurance important?",
-    answer:
-      "Life insurance is important because it provides financial protection for your loved ones in the event of your passing. It helps cover expenses such as funeral costs, outstanding debts, and ongoing living expenses, ensuring your family's financial security.",
-  },
-  {
-    question:
-      "What is the difference between term life and whole life insurance?",
-    answer:
-      "Term life insurance provides coverage for a specified period, while whole life insurance offers lifelong coverage with a cash value component. Term life is generally more affordable, whereas whole life builds cash value over time.",
-  },
-  {
-    question: "Can I borrow against my life insurance policy?",
-    answer:
-      "Yes, you can borrow against the cash value of a whole life or universal life insurance policy. The loan amount, plus interest, will reduce the death benefit if not repaid.",
-  },
-  {
-    question: "What factors affect life insurance premiums?",
-    answer:
-      "Factors affecting life insurance premiums include age, health, lifestyle, occupation, coverage amount, and policy type. Non-smokers and individuals in good health typically pay lower premiums.",
-  },
-]
-
-export const generalLiabilityFAQ: FAQ[] = [
-  {
-    question: "What does general liability insurance cover?",
-    answer:
-      "General liability insurance covers claims of bodily injury, property damage, and personal injury arising from your business operations. It provides protection against legal costs and damages awarded in lawsuits.",
-  },
-  {
-    question: "How much does general liability insurance cost?",
-    answer:
-      "The cost of general liability insurance varies based on factors such as industry, location, coverage limits, and number of employees. On average, small businesses can expect to pay between $400 and $1,500 annually.",
-  },
-  {
-    question: "Why is general liability insurance important for businesses?",
-    answer:
-      "General liability insurance is important for businesses because it protects against financial loss due to claims of bodily injury, property damage, and personal injury. It helps cover legal defense costs and damages, ensuring your business remains financially stable.",
-  },
-  {
-    question:
-      "What is the difference between general liability and professional liability insurance?",
-    answer:
-      "General liability insurance covers bodily injury, property damage, and personal injury claims, while professional liability insurance (also known as errors and omissions insurance) covers claims related to professional services and advice.",
-  },
-  {
-    question: "Can general liability insurance cover product liability?",
-    answer:
-      "Yes, general liability insurance can include product liability coverage, which protects your business against claims of injury or damage caused by products you manufacture, sell, or distribute.",
-  },
-  {
-    question: "How do I file a general liability insurance claim?",
-    answer:
-      "To file a general liability insurance claim, contact your insurance provider as soon as possible, provide detailed information about the incident, gather any necessary documentation, and follow your insurer's claims process.",
-  },
-]
-
-export const umbrellaFAQ: FAQ[] = [
-  {
-    question: "What does umbrella insurance cover?",
-    answer:
-      "Umbrella insurance provides additional liability coverage beyond the limits of your existing policies, such as homeowners, auto, or renters insurance. It covers a wide range of risks, including bodily injury, property damage, and personal liability claims.",
-  },
-  {
-    question: "How much does umbrella insurance cost?",
-    answer:
-      "The cost of umbrella insurance varies based on factors such as coverage limits and existing policies. On average, umbrella insurance can cost between $150 and $300 per year for $1 million in additional coverage.",
-  },
-  {
-    question: "Why do I need umbrella insurance?",
-    answer:
-      "Umbrella insurance provides an extra layer of protection against significant claims and lawsuits. It helps safeguard your assets and financial future, ensuring you are covered beyond the limits of your primary insurance policies.",
-  },
-  {
-    question:
-      "What is the difference between umbrella insurance and excess liability insurance?",
-    answer:
-      "Umbrella insurance provides broader coverage, including risks not covered by underlying policies, while excess liability insurance only increases the coverage limits of existing liability policies without adding new types of coverage.",
-  },
-  {
-    question: "Can umbrella insurance cover legal defense costs?",
-    answer:
-      "Yes, umbrella insurance can cover legal defense costs associated with liability claims that exceed the limits of your primary insurance policies. This includes attorney fees, court costs, and settlement expenses.",
-  },
-  {
-    question:
-      "Do I need umbrella insurance if I already have high liability limits on my primary policies?",
-    answer:
-      "Even if you have high liability limits on your primary policies, umbrella insurance provides additional coverage and peace of mind. It can protect you from large claims that exceed your primary policy limits and offer broader coverage for various liability risks.",
-  },
-]
