+++ config/services/health.ts
@@ -1,0 +1,7 @@
+import { Metadata } from "next"
+
+export const autoMetaData: Metadata = {
+  title: "",
+  description: "",
+  keywords: [],
+}
