+++ config/site.ts
@@ -1,89 +1,6 @@
-import { SiteConfig, Location, Address, Hours } from "types"
+import { SiteConfig, Hours } from "types"
 import { env } from "@/env.mjs"
-
-export const defaultHours: Hours = {
-  Sunday: null,
-  Monday: { start: "08:30 AM", end: "05:30 PM" },
-  Tuesday: { start: "08:30 AM", end: "05:30 PM" },
-  Wednesday: { start: "08:30 AM", end: "05:30 PM" },
-  Thursday: { start: "08:30 AM", end: "05:30 PM" },
-  Friday: { start: "08:30 AM", end: "05:30 PM" },
-  Saturday: null,
-}
-
-const defaultHoursShortened = {
-  weekdays: "",
-  weekends: "",
-}
-
-const collegeParkAddress: Address = {
-  street: "6900 Wisconsin Ave ste 450",
-  city: "Bethesda",
-  state: "MD",
-  zip: "20815",
-}
-
-const bethesdaAddress: Address = {
-  street: "6900 Wisconsin Ave ste 450",
-  city: "Bethesda",
-  state: "MD",
-  zip: "20815",
-}
-
-const fredricksburgAddress: Address = {
-  street: "6900 Wisconsin Ave ste 450",
-  city: "Bethesda",
-  state: "MD",
-  zip: "20815",
-}
-
-const collegeParkLocation: Location = {
-  name: "Main Office",
-  googleMapsURL: "https://maps.app.goo.gl/LpgcXg2o7WFD39JR6",
-  hours: defaultHours,
-  hoursShortened: defaultHoursShortened,
-  address: collegeParkAddress,
-  getAddress: function () {
-    const address = this.address as Address
-    const lineOne = address.street
-    const lineTwo = `${address.city}, ${address.state} ${address.zip}`
-    return { lineOne, lineTwo }
-  },
-}
-
-const bethesdaLocation: Location = {
-  name: "Bethesda, MD -  Office",
-  googleMapsURL: "https://maps.app.goo.gl/LpgcXg2o7WFD39JR6",
-  hours: defaultHours,
-  hoursShortened: defaultHoursShortened,
-  address: bethesdaAddress,
-  getAddress: function () {
-    const address = this.address as Address
-    const lineOne = address.street
-    const lineTwo = `${address.city}, ${address.state} ${address.zip}`
-    return { lineOne, lineTwo }
-  },
-}
-
-const fredricksburgLocation: Location = {
-  name: "Fredricksburg, VA - Office",
-  googleMapsURL: "https://maps.app.goo.gl/LpgcXg2o7WFD39JR6",
-  hours: defaultHours,
-  hoursShortened: defaultHoursShortened,
-  address: fredricksburgAddress,
-  getAddress: function () {
-    const address = this.address as Address
-    const lineOne = address.street
-    const lineTwo = `${address.city}, ${address.state} ${address.zip}`
-    return { lineOne, lineTwo }
-  },
-}
-
-export const locations: Location[] = [
-  collegeParkLocation,
-  bethesdaLocation,
-  fredricksburgLocation,
-]
+import { locations } from "./locations"
 
 export const siteConfig: SiteConfig = {
   name: "Kamil Agency",
